cms-api/requirements.txt
asgiref==3.2.10
astroid==2.4.2
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.0
django-log-to-telegram==2020.1.23
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-core==1.4.3
google-cloud-storage==1.33.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
django-cors-headers==3.6.0
googleapis-common-protos==1.52.0
idna==2.10
inflection==0.5.1
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==2.0.3
numpy==1.19.1
packaging==20.4
pandas==1.1.1
passlib==1.7.1
protobuf==3.14.0
psycopg2-binary==2.8.4
ptvsd==4.3.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
rsa==4.6
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
SQLAlchemy==1.3.9
sqlparse==0.3.1
timestring==1.6.4
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
wrapt==1.12.1
xlrd==1.2.0
elasticsearch==7.11.0
